reckon ammendments
