+++ scripts/util.py
@@ -1,0 +1,108 @@
+from subprocess import call, Popen, run
+import shlex
+import itertools as it
+import uuid
+from datetime import datetime
+import json
+import os
+
+from typing import Dict, Any, AnyStr
+
+import math
+
+def call_tcp_dump(tcpdump_path, cmd):
+    tcp_dump_cmd = [
+        "tcpdump",
+        "-i",
+        "any",
+        "-w",
+        tcpdump_path,
+        "net",
+        "10.0.0.0/16",
+        "-n",
+    ]
+    print(tcp_dump_cmd)
+    p = Popen(tcp_dump_cmd)
+    call(cmd)
+    p.terminate()
+
+default_parameters = {
+        'system':'etcd',
+        'client':'go',
+        'topo':'simple',
+        'failure':'none',
+        'nn':3,
+        'nc':1,
+        'delay':20,
+        'loss':0,
+        'jitter':0,
+        'ncpr':'False',
+        'mtbf':1,
+        'kill_n':0,
+        'write_ratio':1,
+        'rate':1000,
+        'duration':60,
+        'tag':'tag',
+        'tcpdump':False,
+        'arrival_process':'uniform',
+        'repeat':-1,
+        'failure_timeout':1,
+        'delay_interval':0.1,
+        'notes':{},
+        }
+
+def run_test(folder_path, config : Dict[str, Any]):
+    run('rm -rf /data/*', shell=True).check_returncode()
+    run('mn -c', shell=True).check_returncode()
+    run('pkill client', shell=True)
+
+    uid = uuid.uuid4()
+
+    for k,_ in config.items():
+        assert(k in default_parameters.keys())
+
+    # Set params
+    params = default_parameters.copy()
+    for k,v in config.items():
+        params[k] = v
+    del config
+
+    print(params)
+
+    assert (params['repeat'] != -1)
+
+    result_folder = f"{folder_path}/{uid}/"
+    log_path    = result_folder + f"logs"
+    config_path = result_folder + f"config.json"
+    result_path = result_folder + f"res.json"
+    tcpdump_path= result_folder + f"tcpdump.pcap"
+
+    cmd = " ".join([
+        f"python -m reckon {params['system']} {params['topo']} {params['failure']}",
+        f"--number-nodes {params['nn']} --number-clients {params['nc']} --client {params['client']}",
+        f"--link-latency {params['delay']} --link-loss {params['loss']} --link-jitter {params['jitter']}",
+        f"--new_client_per_request {params['ncpr']}",
+        f"--mtbf {params['mtbf']} --kill-n {params['kill_n']}",
+        f"--write-ratio {params['write_ratio']}",
+        f"--rate {params['rate']} --duration {params['duration']}",
+        f"--arrival-process {params['arrival_process']}",
+        f"--system_logs {log_path} --result-location {result_path} --data-dir=/data",
+        f"--failure_timeout {params['failure_timeout']}",
+        f"--delay_interval {params['delay_interval']}",
+        ])
+
+    run(f'mkdir -p {result_folder}', shell=True).check_returncode()
+    run(f'mkdir -p {log_path}', shell=True).check_returncode()
+
+    with open(config_path, "w") as of:
+        json.dump(params, of)
+
+    print(f"RUNNING TEST")
+    print(cmd)
+
+    cmd = shlex.split(cmd)
+
+    if params['tcpdump']:
+        call_tcp_dump(tcpdump_path, cmd)
+    else:
+        call(cmd)
